I have implemented crude input control, which detects if the left arrow key is being depressed and then moves the avatar to the left when it is.
Unfortunately due to the sheer quantity of tiles being rendered on the screen at a single time, this movement is so far undetectable, as the FPS of the game is way way down.

I will need to come up with a solution to either combine together tiles to make it easier for them to be rendered, or come up with a solution which best fits the minimum number of tiles.